Clean up main.js.
