+++ src/ViewModels/ApplicationViewModel.js
@@ -1,15 +1,23 @@
 "use strict";
 
+/*global require, URI*/
 var CesiumEvent = require('../../third_party/cesium/Source/Core/Event');
 var Clock = require('../../third_party/cesium/Source/Core/Clock');
 var DataSourceCollection = require('../../third_party/cesium/Source/DataSources/DataSourceCollection');
+var defaultValue = require('../../third_party/cesium/Source/Core/defaultValue');
+var defined = require('../../third_party/cesium/Source/Core/defined');
+var FeatureDetection = require('../../third_party/cesium/Source/Core/FeatureDetection');
 var knockout = require('../../third_party/cesium/Source/ThirdParty/knockout');
+var loadJson = require('../../third_party/cesium/Source/Core/loadJson');
+var Rectangle = require('../../third_party/cesium/Source/Core/Rectangle');
+var when = require('../../third_party/cesium/Source/ThirdParty/when');
 
 var CatalogViewModel = require('./CatalogViewModel');
 var corsProxy = require('../Core/corsProxy');
 var NowViewingViewModel = require('./NowViewingViewModel');
 var ServicesViewModel = require('./ServicesViewModel');
 var ViewerMode = require('./ViewerMode');
+var ViewModelError = require('./ViewModelError');
 
 /**
  * The overall view-model for National Map.
@@ -54,6 +62,12 @@ var ApplicationViewModel = function() {
      * @type {Clock}
      */
     this.clock = new Clock();
+
+    /**
+     * Gets or sets the initial bounding box of the camera's view.
+     * @type {Rectangle}
+     */
+    this.initialBoundingBox = Rectangle.MAX_VALUE;
 
     /**
      * Gets or sets the {@link corsProxy} used to determine if a URL needs to be proxied and to proxy it if necessary.
@@ -103,5 +117,141 @@ var ApplicationViewModel = function() {
 
     knockout.track(this, ['viewerMode']);
 };
+
+/**
+ * Starts the application.
+ *
+ * @param {Object} options Object with the following properties:
+ * @param {String} [options.applicationUrl] The URL of the application.  Typically this is obtained from window.location.  This URL, if
+ *                                          supplied, is parsed for startup parameters.
+ * @param {String} [options.configUrl='config.json'] The URL of the file containing configuration information, such as the list of domains to proxy.
+ * @param {String} [options.initializationUrl] The URL of the main file containing initialization information, such as the list of
+ *                                             catalog items.
+ * @param {Boolean} [options.useApplicationUrlHashAsInitSource=true] true to parse the applicationUrl as an init source.  The hash may be of the form
+ *                                                                   'start=???', where ??? is a JSON-encoded initialization object, or it may be
+ *                                                                   a simple string.  If it's a simple string, a file named 'init_' + hash + '.json'
+ *                                                                   will be loaded as the init source.  For example, #vic will load init_vic.json.
+ */
+ApplicationViewModel.prototype.start = function(options) {
+    var applicationUrl = defaultValue(options.applicationUrl, '');
+    var configUrl = defaultValue(options.configUrl, 'config.json');
+
+    var uri = new URI(applicationUrl);
+    var urlParameters = uri.search(true);
+    var hash = uri.fragment();
+
+    var that = this;
+    return loadJson(options.configUrl).then(function(config) {
+        if (defined(options.initializationUrl)) {
+            that.initSources.push(options.initializationUrl);
+        }
+
+        // If the URL includes a hash, try loading the corresponding init file.
+        var startData = {};
+        if (defaultValue(options.useApplicationUrlHashAsInitSource, true)) {
+            if (defined(hash) && hash.length > 0) {
+                if (hash.indexOf('start=') === 0) {
+                    startData = JSON.parse(decodeURIComponent(hash.substring(6)));
+                } else if (hash.toLowerCase() !== 'populate-cache') {
+                    that.initSources.push('init_' + hash + ".json");
+                }
+            }
+        }
+
+        var initSources = that.initSources.slice();
+
+        // Include any initSources specified in the URL.
+        if (defined(startData.initSources)) {
+            for (var i = 0; i < startData.initSources.length; ++i) {
+                var initSource = startData.initSources[i];
+                if (initSources.indexOf(initSource) < 0) {
+                    initSources.push(initSource);
+
+                    // Only add external files to the application's list of init sources.
+                    if (typeof initSource === 'string') {
+                        that.initSources.push(initSource);
+                    }
+                }
+            }
+        }
+
+        // Load all of the init sources.
+        return when.all(initSources.map(loadInitSource), function(initSources) {
+            var corsDomains = [];
+            var i;
+            var initSource;
+
+            for (i = 0; i < initSources.length; ++i) {
+                initSource = initSources[i];
+                if (!defined(initSource)) {
+                    continue;
+                }
+
+                // Extract the list of CORS-ready domains from the init sources.
+                if (defined(initSource.corsDomains)) {
+                    corsDomains.push.apply(corsDomains, initSource.corsDomains);
+                }
+
+                // The last init source to specify a camera position wins.
+                if (defined(initSource.camera)) {
+                    that.initialBoundingBox = initSource.camera;
+                }
+            }
+
+            // Configure the proxy.
+            // IE versions prior to 10 don't support CORS, so always use the proxy.
+            var alwaysUseProxy = (FeatureDetection.isInternetExplorer() && FeatureDetection.internetExplorerVersion()[0] < 10);
+            corsProxy.setProxyList(config.proxyDomains, corsDomains, alwaysUseProxy);
+
+            var promises = [];
+
+            // Make another pass over the init sources to update the catalog and load services.
+            for (i = 0; i < initSources.length; ++i) {
+                initSource = initSources[i];
+                if (!defined(initSource)) {
+                    continue;
+                }
+
+                if (defined(initSource.catalog)) {
+                    var isUserSupplied;
+                    if (initSource.isFromExternalFile) {
+                        isUserSupplied = false;
+                    } else if (initSource.catalogOnlyUpdatesExistingItems) {
+                        isUserSupplied = undefined;
+                    } else {
+                        isUserSupplied = true;
+                    }
+
+                    promises.push(that.catalog.updateFromJson(initSource.catalog, {
+                        onlyUpdateExistingItems: initSource.catalogOnlyUpdatesExistingItems,
+                        isUserSupplied: isUserSupplied
+                    }));
+                }
+
+                if (defined(initSource.services)) {
+                    that.services.services.push.apply(that.services, initSource.services);
+                }
+            }
+
+            return when.all(promises);
+        });
+    });
+};
+
+function loadInitSource(source) {
+    if (typeof source === 'string') {
+        return loadJson(source).then(function(initSource) {
+            initSource.isFromExternalFile = true;
+            return initSource;
+        }).otherwise(function() {
+            throw new ViewModelError({
+                title: 'Error loading initialization source',
+                message: 'An error occurred while loading initialization information from ' + source + '.  This may indicate that you followed an invalid link or that there is a problem with your Internet connection.'
+            });
+        });
+    } else {
+        return source;
+    }
+}
 
 module.exports = ApplicationViewModel;
